behavioral-backend/requirements.txt
# Core NLP
textblob>=0.17.1

# Upstash Redis client
upstash-redis>=0.15.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
numpy>=1.24.0
pytz>=2023.3
